# carbon_foot_print/documentation/swagger/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:5000
    description: Generated server url
paths:
  /api/user/volunteers:
    get:
      tags:
        - user-volunteer-rest-controller
      operationId: listVolunteers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserVolunteerDTOResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - user-volunteer-rest-controller
      operationId: updateUserVolunteer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserVolunteerDTORequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - user-volunteer-rest-controller
      operationId: saveVolunteer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserVolunteerDTORequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/categories:
    get:
      tags:
        - category-rest-controller
      operationId: listAll
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDTOResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - category-rest-controller
      operationId: updateUserVolunteer_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryWithIdDTORequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - category-rest-controller
      operationId: saveCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTORequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/categories/ids:
    post:
      tags:
        - category-rest-controller
      operationId: getFullCategories
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryByIds'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
  /api/categories/enabled:
    get:
      tags:
        - category-rest-controller
      operationId: getCategoriesEnabled
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDTOResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - category-rest-controller
      operationId: getFullEnabledCategories
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryByIds'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/categories/{categoryId}:
    get:
      tags:
        - category-rest-controller
      operationId: getCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDTOResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
  /api/categories/name:
    get:
      tags:
        - category-rest-controller
      operationId: getCategoryBy
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryDTOResponse'
        '400':
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    PermissionDTORequest:
      required:
        - description
        - name
      type: object
      properties:
        idPermission:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    RoleWithIdDTORequest:
      required:
        - state
        - typeRole
      type: object
      properties:
        idRole:
          type: integer
          format: int64
        typeRole:
          type: string
        state:
          type: boolean
        permissions:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/PermissionDTORequest'
    UserVolunteerDTORequest:
      required:
        - documentType
        - lastNames
        - names
        - password
        - personalEmail
        - position
        - state
        - username
      type: object
      properties:
        documentNumber:
          type: integer
          format: int64
        documentType:
          maxLength: 45
          minLength: 6
          type: string
        names:
          maxLength: 60
          minLength: 2
          type: string
        lastNames:
          maxLength: 60
          minLength: 2
          type: string
        personalPhone:
          pattern: '[3][0-9]{9}'
          type: string
        personalEmail:
          type: string
        username:
          type: string
        password:
          type: string
        roles:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/RoleWithIdDTORequest'
        state:
          type: boolean
        position:
          type: string
...
    SourceWithIdDTORequest:
      required:
        - sourceName
      type: object
      properties:
        idSource:
          type: integer
          format: int64
        sourceName:
          type: string
    PollutionTypeWithIdDTORequest:
      required:
        - pollutionSources
        - pollutionTypeDescription
        - pollutionTypeEmissionFactor
        - pollutionTypeId
        - pollutionTypeName
        - pollutionTypeUnits
      type: object
      properties:
        pollutionTypeId:
          type: integer
          format: int64
        pollutionTypeName:
          type: string
        pollutionTypeDescription:
          type: string
        pollutionTypeEmissionFactor:
          type: number
          format: double
        pollutionTypeUnits:
          type: string
        pollutionSources:
          type: array
          items:
            $ref: '#/components/schemas/SourceWithIdDTORequest'
    CategoryWithIdDTORequest:
      required:
        - categoryDescription
        - categoryName
        - categoryScope
        - categoryStatus
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
        categoryName:
          maxLength: 50
          minLength: 2
          type: string
        categoryDescription:
          type: string
        categoryScope:
          type: string
        categoryStatus:
          type: boolean
        categoryPollution:
          type: array
          items:
            $ref: '#/components/schemas/PollutionTypeWithIdDTORequest'
...
    CategoryDTORequest:
      required:
        - categoryDescription
        - categoryName
        - categoryScope
        - categoryStatus
      type: object
      properties:
        categoryName:
          maxLength: 50
          minLength: 2
          type: string
        categoryDescription:
          type: string
        categoryScope:
          type: string
        categoryStatus:
          type: boolean
        categoryPollution:
          type: array
          items:
            $ref: '#/components/schemas/PollutionTypeWithIdDTORequest'
    CategoryByIds:
      required:
        - ids
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
            format: int64
...
    PermissionDTOResponse:
      type: object
      properties:
        idPermission:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    RoleDTOResponse:
      type: object
      properties:
        idRole:
          type: integer
          format: int64
        typeRole:
          type: string
        state:
          type: boolean
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionDTOResponse'
    UserVolunteerDTOResponse:
      type: object
      properties:
        documentNumber:
          type: integer
          format: int64
        documentType:
          type: string
        names:
          type: string
        lastNames:
          type: string
        personalPhone:
          type: string
        personalEmail:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDTOResponse'
        state:
          type: boolean
        position:
          type: string
...
    CategoryDTOResponse:
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        categoryDescription:
          type: string
        categoryScope:
          type: string
        categoryStatus:
          type: boolean
    Error:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        httpCode:
          type: integer
          format: int64
        url:
          type: string
        method:
          type: string